Move gaphas.solver to a package
Separate the classes in their own files.
