+++ gaphas/solver/__init__.py
@@ -1,0 +1,11 @@
+from gaphas.solver.projection import Projection
+from gaphas.solver.solver import Solver, solvable
+from gaphas.solver.variable import (
+    NORMAL,
+    REQUIRED,
+    STRONG,
+    VERY_STRONG,
+    VERY_WEAK,
+    WEAK,
+    Variable,
+)
